Dirty fix to rare (but triggerable) race condition in OnResize()
A segfault occurs when the wxEVENT_SIZE event is triggered and the
window the event occurs on has not shown up yet. This seems to be a
race condition, since it's very tricky to reproduce this bug. We
were able to reproduce it by starting fresh (e.g. removing
~/.Wabbitemu), opening an existing ROM and pressing ENTER to
confirm, and then clicking Finish.

We are unsure whether this bug is caused by an error in wxWabbitemu
or a bug in wxWidgets. Regardless, this is temporarily fixed by using
a variable to determine whether the window has shown up or not via
the wxEVENT_SHOW event, and then checking this event within OnResize().

This should fix issue #1.