Removed download option/funcs + cleaned up wizard code
TI no longer allows you to automatically download the OS. Instead, you
must go to their website directly to download the OS. Upsteam attempted
to fix the issue in Feb 2016, but as of Aug 2016 has declared the issue
unfixable, and suggests to download directly as well. Therefore, we are
following suit and removing the feature completely, since there doesn't
seem to be any way around this. We replaced this option with a button
to open TI's download page so users can grab the OS by themselves.

Since the change involved removing some core options, we also took the
time to clean up the wizard handling as well, resulting in better,
hopefully more stable code.

Other changes bundled in this commit:

 * Move all init + external label/button control to romwizard.cpp
   (especially from wizardstart.cpp)
 * Created UpdateWizard in wizardos.cpp for updating from romwizard.cpp
 * Corrected checking cases to correctly enable/disable forward button
   in wizardos.cpp
 * Add error message when incorrect file type is selected in
   wizardos.cpp
 * Optimized path correctness checking when field is empty in
   wizardos.cpp
 * Add hack to prevent wizard page changing if buttons are disabled in
   romwizard.cpp
 * Handle not selecting a file path to save to in romwizard.cpp
 * Remove useless veto in romwizard.cpp for OnFinished (doesn't work)