Fix corner case of Next btn when invalid OS specified
When reaching the OS file selection page in the wizard, the user could
specify an invalid OS file, then click Back. Once they click Back, the
Next button is still disabled due to the disabling of the button on
the OS selection page. This is fixed by always enabling the Next button
on the calculator type page within romwizard.cpp.