Fix corner case of Back/Next error msg when invalid OS specified
When reaching the OS file selection page in the wizard, the user could
specify an invalid OS file, get an error message, then click Back. Once
they click Back, they could click Next again, only to receive the same
error message again. This is fixed by adding an optional argument to
UpdateWizard to determine whether an error message should be shown or
not, and only enabling the error message when a file is selected.